renamed Player to Lobby. Updated import/exportsfor actions/reducers.
