+++ src/actions/index.js
@@ -1,6 +1,6 @@
 export { toggleModal, updateTeamSettings } from './appActions';
 export type { ModalAction, LoadingAction, TeamSettings } from './appActions';
-export { playersCountIncrement, playersCountDecrement, updatePlayersList } from './playersActions';
-export type { CountIncrementAction, CountDecrementAction, updatePlayersListAction, Player, PlayersList } from './playersActions';
+export { playersCountIncrement, playersCountDecrement, updatePlayersList } from './lobbyActions';
+export type { CountIncrementAction, CountDecrementAction, updatePlayersListAction, Player, PlayersList } from './lobbyActions';
 export {  updateQueueList, returnPlayerToLobby, queueCountDecrement, queueCountIncrement} from './queueActions';
 export type { QueueList, ReturnPlayerToLobby, QueueCountDecrement, QueueCountIncrement, MovePlayerToTeam } from './queueActions';
