+++ src/components/MainArea.js
@@ -4,13 +4,13 @@ import React from 'react';
 import '../styles/MainArea.css'
 import QueueArea from './QueueArea';
 import TeamsArea from './TeamsArea';
-import PlayersArea from './PlayersArea';
+import LobbyArea from './LobbyArea';
 
 
 function MainArea() {
   return (
     <div className="MainArea-container">
-      <PlayersArea />
+      <LobbyArea />
       <QueueArea />
       <TeamsArea />
     </div>
