+++ src/constants/action-types.js
@@ -22,7 +22,7 @@ export {
   PLAYERS_COUNT_INCREMENT,
   PLAYERS_COUNT_DECREMENT,
   UPDATE_PLAYERS_LIST,
-  INITIALIZE_PLAYERS_LIST
+  INITIALIZE_PLAYERS_LIST,
   UPDATE_QUEUE_LIST,
   RETURN_PLAYER_TO_LOBBY,
   MOVE_PLAYER_TO_TEAM,
