+++ src/reducers/index.js
@@ -1,13 +1,13 @@
 import { combineReducers } from 'redux';
 
 import appReducers from './appReducers';
-import playersReducers from './playersReducers';
+import lobbyReducers from './lobbyReducers';
 import queueReducers from './queueReducers';
 
 
 export const rootReducer = combineReducers({
   app: appReducers,
-  players: playersReducers,
+  lobby: lobbyReducers,
   queue: queueReducers
 });
